fix: support for proper identification of '.' remote paths in `reference::remote::Name` (#450)
